docs(numeric): update documentation stories
